355-rename-lagom-dev-mode-components: rollback breaking changed in the public API and move the registry client in the devmode package
