+++ dev/service-registry/client-core/src/main/scala/com/lightbend/lagom/devmode/internal/registry/AbstractLoggingServiceRegistryClient.scala
@@ -2,7 +2,7 @@
  * Copyright (C) Lightbend Inc. <https://www.lightbend.com>
  */
 
-package com.lightbend.lagom.internal.registry
+package com.lightbend.lagom.devmode.internal.registry
 
 import java.net.URI
