+++ docs/manual/scala/guide/cluster/code/docs/home/scaladsl/persistence/JdbcBlogApplicationLoader.scala
@@ -6,7 +6,7 @@ package docs.home.scaladsl.persistence
 
 import com.lightbend.lagom.scaladsl.api.ServiceLocator
 import com.lightbend.lagom.scaladsl.api.ServiceLocator.NoServiceLocator
-import com.lightbend.lagom.devmode.scaladsl.LagomDevModeComponents
+import com.lightbend.lagom.scaladsl.devmode.LagomDevModeComponents
 import com.lightbend.lagom.scaladsl.persistence.jdbc.JdbcPersistenceComponents
 import com.lightbend.lagom.scaladsl.server._
 import com.softwaremill.macwire._
