+++ docs/manual/scala/guide/cluster/code/docs/home/scaladsl/persistence/ShoppingCartLoader.scala
@@ -10,7 +10,7 @@ import com.lightbend.lagom.scaladsl.api.Descriptor
 import com.lightbend.lagom.scaladsl.api.ServiceCall
 import com.lightbend.lagom.scaladsl.api.transport.Method
 
-import com.lightbend.lagom.devmode.scaladsl.LagomDevModeComponents
+import com.lightbend.lagom.scaladsl.devmode.LagomDevModeComponents
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
 import com.lightbend.lagom.scaladsl.broker.kafka.LagomKafkaComponents
 import com.lightbend.lagom.scaladsl.akka.discovery.AkkaDiscoveryComponents
