+++ docs/manual/scala/guide/production/code/AkkaDiscoveryIntegration.scala
@@ -8,7 +8,7 @@ package akkadiscoveryservicelocator {
   import docs.scaladsl.services.lagomapplication.HelloApplication
 
   //#akka-discovery-service-locator
-  import com.lightbend.lagom.devmode.scaladsl.LagomDevModeComponents
+  import com.lightbend.lagom.scaladsl.devmode.LagomDevModeComponents
   import com.lightbend.lagom.scaladsl.server._
   import com.lightbend.lagom.scaladsl.akka.discovery.AkkaDiscoveryComponents
